Fix filling x axis values for date histogram (#35247)
The interval parameter is used to fill the gaps between missing buckets. If we are using a calendar interval it not always correspond correctly to his milliseconds representation. E.g. 1y interval can correspond to 365 or 366 days depending on the interval and. If transformed to milliseconds it can get one of those two values depending of the interval start and end. The same happens for months.

Add a new esarchive with 7 years of data, used for those functional tests

Removed angular dependency from zero injection
