+++ src/legacy/ui/public/agg_response/point_series/__tests__/_init_x_axis.js
@@ -18,6 +18,7 @@
  */
 
 import expect from '@kbn/expect';
+import moment from 'moment';
 import { initXAxis } from '../_init_x_axis';
 import { makeFakeXAspect } from '../_fake_x_aspect';
 
@@ -106,15 +107,14 @@ describe('initXAxis', function () {
   });
 
   it('reads the interval param from the x agg', function () {
-    chart.aspects.x[0].params.interval = 10;
+    chart.aspects.x[0].params.interval = 'P1D';
     initXAxis(chart, table);
     expect(chart)
       .to.have.property('xAxisLabel', 'label')
       .and.have.property('xAxisFormat', chart.aspects.x[0].format)
       .and.have.property('ordered');
 
-    expect(chart.ordered)
-      .to.be.an('object')
-      .and.have.property('interval', 10);
+    expect(moment.isDuration(chart.ordered.interval)).to.be(true);
+    expect(chart.ordered.interval.toISOString()).to.eql('P1D');
   });
 });
