+++ src/legacy/ui/public/vislib/components/zero_injection/inject_zeros.js
@@ -17,35 +17,28 @@
  * under the License.
  */
 
-import { VislibComponentsZeroInjectionOrderedXKeysProvider } from './ordered_x_keys';
-import { VislibComponentsZeroInjectionZeroFilledArrayProvider } from './zero_filled_array';
-import { VislibComponentsZeroInjectionZeroFillDataArrayProvider } from './zero_fill_data_array';
+import { orderXValues } from './ordered_x_keys';
+import { createZeroFilledArray } from './zero_filled_array';
+import { zeroFillDataArray } from './zero_fill_data_array';
 
-export function VislibComponentsZeroInjectionInjectZerosProvider(Private) {
+/*
+* A Kibana data object may have multiple series with different array lengths.
+* This proves an impediment to stacking in the visualization library.
+* Therefore, zero values must be injected wherever these arrays do not line up.
+* That is, each array must have the same x values with zeros filled in where the
+* x values were added.
+*
+* This function and its helper functions accepts a Kibana data object
+* and injects zeros where needed.
+*/
 
-  const orderXValues = Private(VislibComponentsZeroInjectionOrderedXKeysProvider);
-  const createZeroFilledArray = Private(VislibComponentsZeroInjectionZeroFilledArrayProvider);
-  const zeroFillDataArray = Private(VislibComponentsZeroInjectionZeroFillDataArrayProvider);
+export function injectZeros(obj, data, orderBucketsBySum = false) {
+  const keys = orderXValues(data, orderBucketsBySum);
 
-  /*
-   * A Kibana data object may have multiple series with different array lengths.
-   * This proves an impediment to stacking in the visualization library.
-   * Therefore, zero values must be injected wherever these arrays do not line up.
-   * That is, each array must have the same x values with zeros filled in where the
-   * x values were added.
-   *
-   * This function and its helper functions accepts a Kibana data object
-   * and injects zeros where needed.
-   */
+  obj.forEach(function (series) {
+    const zeroArray = createZeroFilledArray(keys, series.label);
+    series.values = zeroFillDataArray(zeroArray, series.values);
+  });
 
-  return function (obj, data, orderBucketsBySum = false) {
-    const keys = orderXValues(data, orderBucketsBySum);
-
-    obj.forEach(function (series) {
-      const zeroArray = createZeroFilledArray(keys, series.label);
-      series.values = zeroFillDataArray(zeroArray, series.values);
-    });
-
-    return obj;
-  };
+  return obj;
 }
