+++ src/legacy/ui/public/vislib/components/zero_injection/ordered_x_keys.js
@@ -19,64 +19,59 @@
 
 import _ from 'lodash';
 import moment from 'moment';
-import { VislibComponentsZeroInjectionUniqKeysProvider } from './uniq_keys';
+import { getUniqKeys } from './uniq_keys';
 
-export function VislibComponentsZeroInjectionOrderedXKeysProvider(Private) {
-  const getUniqKeys = Private(VislibComponentsZeroInjectionUniqKeysProvider);
+/*
+  * Accepts a Kibana data object and returns
+  * an array of x axis values.
+  * values sorted by timestamp if isDate and Date Histogram agg
+  * else values sorted by index
+  */
+export function orderXValues(obj, orderBucketsBySum = false) {
+  if (!_.isObject(obj)) {
+    throw new Error('orderXValues expects an object');
+  }
 
-  /*
-   * Accepts a Kibana data object and returns
-   * an array of x axis values.
-   * values sorted by timestamp if isDate and Date Histogram agg
-   * else values sorted by index
-   */
+  const uniqKeys = getUniqKeys(obj);
+  const uniqKeysPairs = [...uniqKeys.entries()];
 
-  return function (obj, orderBucketsBySum = false) {
-    if (!_.isObject(obj)) {
-      throw new Error('OrderedXKeysUtilService expects an object');
-    }
+  const interval = _.get(obj, 'ordered.interval');
+  const dateInterval = moment.isDuration(interval) ? interval : false;
 
-    const uniqKeys = getUniqKeys(obj);
-    const uniqKeysPairs = [...uniqKeys.entries()];
+  return _(uniqKeysPairs)
+    .sortBy(function (d) {
+      if (d[1].isDate || d[1].isOrdered) {
+        return +d[0];
+      }
+      return orderBucketsBySum ? -d[1].sum : d[1].index;
+    })
+    .map(function (d, i, list) {
+      if (!d[1].isNumber) return d[0];
 
-    const interval = _.get(obj, 'ordered.interval');
-    const dateInterval = moment.isDuration(interval) ? interval : false;
+      const val = +d[0];
+      if (interval == null) return val;
 
-    return _(uniqKeysPairs)
-      .sortBy(function (d) {
-        if (d[1].isDate || d[1].isOrdered) {
-          return +d[0];
+      const gapEdge = parseFloat(_.get(list, [i + 1, 0]));
+      if (isNaN(gapEdge)) return val;
+
+      const vals = [];
+      let next = val;
+
+      if (dateInterval) {
+        next = moment(val);
+        while (next < gapEdge) {
+          vals.push(next.valueOf());
+          next.add(dateInterval);
         }
-        return orderBucketsBySum ? -d[1].sum : d[1].index;
-      })
-      .map(function (d, i, list) {
-        if (!d[1].isNumber) return d[0];
-
-        const val = +d[0];
-        if (interval == null) return val;
-
-        const gapEdge = parseFloat(_.get(list, [i + 1, 0]));
-        if (isNaN(gapEdge)) return val;
-
-        const vals = [];
-        let next = val;
-
-        if (dateInterval) {
-          next = moment(val);
-          while (next < gapEdge) {
-            vals.push(next.valueOf());
-            next.add(dateInterval);
-          }
-        } else {
-          while (next < gapEdge) {
-            vals.push(next);
-            next += interval;
-          }
+      } else {
+        while (next < gapEdge) {
+          vals.push(next);
+          next += interval;
         }
+      }
 
-        return vals;
-      })
-      .flatten()
-      .value();
-  };
+      return vals;
+    })
+    .flatten()
+    .value();
 }
