+++ src/legacy/ui/public/vislib/lib/data.js
@@ -19,16 +19,14 @@
 
 import d3 from 'd3';
 import _ from 'lodash';
-import { VislibComponentsZeroInjectionInjectZerosProvider } from '../components/zero_injection/inject_zeros';
-import { VislibComponentsZeroInjectionOrderedXKeysProvider } from '../components/zero_injection/ordered_x_keys';
+import { injectZeros } from '../components/zero_injection/inject_zeros';
+import { orderXValues } from '../components/zero_injection/ordered_x_keys';
 import { VislibComponentsLabelsLabelsProvider } from '../components/labels/labels';
 import { VislibComponentsColorColorProvider } from '../../vis/components/color/color';
 import { getFormat } from '../../visualize/loader/pipeline_helpers/utilities';
 
 export function VislibLibDataProvider(Private) {
 
-  const injectZeros = Private(VislibComponentsZeroInjectionInjectZerosProvider);
-  const orderKeys = Private(VislibComponentsZeroInjectionOrderedXKeysProvider);
   const getLabels = Private(VislibComponentsLabelsLabelsProvider);
   const color = Private(VislibComponentsColorColorProvider);
 
@@ -441,7 +439,7 @@ export function VislibLibDataProvider(Private) {
      * @returns {Array} Array of x axis values
      */
     xValues(orderBucketsBySum = false) {
-      return orderKeys(this.data, orderBucketsBySum);
+      return orderXValues(this.data, orderBucketsBySum);
     }
 
     /**
