+++ src/legacy/ui/public/visualize/loader/pipeline_helpers/build_pipeline.ts
@@ -22,6 +22,7 @@ import { cloneDeep } from 'lodash';
 import { setBounds } from 'ui/agg_types/buckets/date_histogram';
 import { SearchSource } from 'ui/courier';
 import { AggConfig, Vis, VisParams, VisState } from 'ui/vis';
+import moment from 'moment';
 
 interface SchemaFormat {
   id: string;
@@ -387,7 +388,8 @@ export const buildVislibDimensions = async (
     const xAgg = vis.aggs.getResponseAggs()[dimensions.x.accessor];
     if (xAgg.type.name === 'date_histogram') {
       dimensions.x.params.date = true;
-      dimensions.x.params.interval = xAgg.buckets.getInterval().asMilliseconds();
+      const { esUnit, esValue } = xAgg.buckets.getInterval();
+      dimensions.x.params.interval = moment.duration(esValue, esUnit);
       dimensions.x.params.format = xAgg.buckets.getScaledDateFormat();
       dimensions.x.params.bounds = xAgg.buckets.getBounds();
     } else if (xAgg.type.name === 'histogram') {
